requirements.txt
absl-py==0.11.0
anyio==2.0.2
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bleach==3.2.1
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
entrypoints==0.3
flatbuffers==1.12
h5py==3.1.0
idna==2.10
importlib-metadata==3.3.0
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
jax==0.2.3
jaxlib==0.1.56
jedi==0.17.2
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.10
jupyter-core==4.7.0
jupyter-server==1.2.0
jupyterlab==3.0.1
jupyterlab-pygments==0.1.2
jupyterlab-server==2.1.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
multiprocess==0.70.11.1
nbclassic==0.2.5
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.6
numpy==1.19.5
numpyro==0.4.1
opt-einsum==3.3.0
packaging==20.8
pandas==1.2.0
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pluggy==0.13.1
prometheus-client==0.9.0
prompt-toolkit==3.0.10
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.7.3
pyparsing==2.4.7
pyro-api==0.1.2
pyro-ppl==1.5.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytz==2020.5
pyzmq==20.0.0
requests==2.25.1
scipy==1.6.0
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
sniffio==1.2.0
terminado==0.9.2
testpath==0.4.4
torch==1.7.1
torchaudio==0.7.2
torchvision==0.8.2
tornado==6.1
tqdm==4.55.1
traitlets==5.0.5
typing-extensions==3.7.4.3
ujson==4.0.1
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.4.0
